Get questionnaire data from a parent element
