+++ components/Questionnaire/Questionnaire.js
@@ -1,0 +1,269 @@
+import React, { Component } from 'react'
+import Frame from '../Frame'
+import Loader from '../Loader'
+
+import { css } from 'glamor'
+import { compose } from 'react-apollo'
+import CheckCircle from 'react-icons/lib/md/check-circle'
+
+import {
+  colors,
+  Interaction,
+  mediaQueries,
+  Button,
+  A,
+  InlineSpinner,
+  fontFamilies,
+  RawHtml
+} from '@project-r/styleguide'
+
+import { HEADER_HEIGHT, HEADER_HEIGHT_MOBILE } from '../constants'
+import withT from '../../lib/withT'
+import withAuthorization from '../Auth/withAuthorization'
+import { errorToString } from '../../lib/utils/errors'
+import { Router } from '../../lib/routes'
+import StatusError from '../StatusError'
+import { withQuestionnaireMutation, withQuestionnaireReset } from './enhancers'
+import Questions from './Questions'
+import QuestionnaireClosed from './QuestionnaireClosed'
+
+const { Headline, P } = Interaction
+
+const styles = {
+  intro: css({
+    marginTop: 35
+  }),
+  count: css({
+    background: '#fff',
+    zIndex: 10,
+    position: 'sticky',
+    padding: '10px 0',
+    borderBottom: `0.5px solid ${colors.divider}`,
+    display: 'flex',
+    minHeight: 55,
+    top: HEADER_HEIGHT - 1,
+    [mediaQueries.onlyS]: {
+      top: HEADER_HEIGHT_MOBILE - 1
+    }
+  }),
+  actions: css({
+    textAlign: 'center',
+    margin: '20px auto 20px auto'
+  }),
+  reset: css({
+    textAlign: 'center',
+    marginTop: 10
+  }),
+  strong: css({
+    fontFamily: fontFamilies.sansSerifMedium
+  }),
+  error: css({
+    color: colors.error,
+    fontFamily: fontFamilies.sansSerifMedium
+  }),
+  closed: css({
+    marginTop: 35,
+    background: colors.primaryBg,
+    width: '100%',
+    alignItems: 'center',
+    justifyContent: 'center',
+    padding: 30,
+    textAlign: 'center',
+    marginBottom: 30
+  }),
+  progressIcon: css({
+    marginLeft: 5,
+    marginTop: 3,
+    minHeight: 30
+  })
+}
+
+class Page extends Component {
+  constructor(props) {
+    super(props)
+    this.state = {}
+  }
+
+  processSubmit = (fn, ...args) => {
+    this.setState({ updating: true })
+    return fn(...args)
+      .then(() =>
+        this.setState(() => ({
+          updating: false,
+          error: null
+        }))
+      )
+      .catch(error => {
+        this.setState(() => ({
+          updating: false,
+          submitting: false,
+          error
+        }))
+      })
+  }
+
+  handleSubmit = () => {
+    this.setState({ submitting: true })
+    const {
+      submitQuestionnaire,
+      data: {
+        questionnaire: { id }
+      }
+    } = this.props
+    this.processSubmit(submitQuestionnaire, id).then(() =>
+      Router.pushRoute('/verlag').then(() => window.scrollTo(0, 0))
+    )
+  }
+
+  handleReset = e => {
+    const {
+      resetQuestionnaire,
+      data: {
+        questionnaire: { id }
+      }
+    } = this.props
+    e.preventDefault()
+    this.processSubmit(resetQuestionnaire, id).then(() => window.scrollTo(0, 0))
+  }
+
+  render() {
+    const {
+      data,
+      t,
+      meta,
+      showResults,
+      pageClosed,
+      questionnaireName,
+      externalSubmit,
+      hideCount
+    } = this.props
+
+    return (
+      <Frame meta={meta}>
+        <Loader
+          loading={data.loading}
+          error={data.error}
+          render={() => {
+            const now = new Date()
+            // handle not found or not started
+            if (
+              !data.questionnaire ||
+              new Date(data.questionnaire.beginDate) > now
+            ) {
+              return (
+                <StatusError
+                  statusCode={404}
+                  serverContext={this.props.serverContext}
+                />
+              )
+            }
+
+            const hasEnded = now > new Date(data.questionnaire.endDate)
+
+            // handle already submitted
+            const {
+              questionnaire: { userHasSubmitted, questions }
+            } = data
+            const { error, submitting, updating } = this.state
+            if (!submitting && (hasEnded || userHasSubmitted)) {
+              return pageClosed ? (
+                pageClosed
+              ) : (
+                <QuestionnaireClosed
+                  submitted={userHasSubmitted}
+                  showResults={showResults}
+                />
+              )
+            }
+
+            // handle questions
+            const questionCount = questions.filter(Boolean).length
+            const userAnswerCount = questions
+              .map(q => q.userAnswer)
+              .filter(Boolean).length
+            const questionnairePath = questionnaireName
+              ? `/${questionnaireName}/`
+              : '/'
+            return (
+              <div>
+                <Headline>
+                  {t(`questionnaire${questionnairePath}title`)}
+                </Headline>
+                <div {...styles.intro}>
+                  <RawHtml
+                    type={P}
+                    dangerouslySetInnerHTML={{
+                      __html: t(`questionnaire${questionnairePath}intro`)
+                    }}
+                  />
+                  <br />
+                </div>
+                {!hideCount && (
+                  <div {...styles.count}>
+                    {error ? (
+                      <P {...styles.error}>{errorToString(error)}</P>
+                    ) : (
+                      <>
+                        <P {...styles.strong}>
+                          {t('questionnaire/header', {
+                            questionCount,
+                            userAnswerCount
+                          })}
+                        </P>
+                        {questionCount === userAnswerCount ? (
+                          <div {...styles.progressIcon}>
+                            <CheckCircle size={22} color={colors.primary} />
+                          </div>
+                        ) : updating || submitting ? (
+                          <div style={{ marginLeft: 5, marginTop: 3 }}>
+                            <InlineSpinner size={24} />
+                          </div>
+                        ) : null}
+                      </>
+                    )}
+                  </div>
+                )}
+                <Questions
+                  questions={questions}
+                  disabled={userHasSubmitted}
+                  processSubmit={this.processSubmit}
+                />
+                {!externalSubmit && (
+                  <div {...styles.actions}>
+                    <Button
+                      primary
+                      onClick={this.handleSubmit}
+                      disabled={updating || submitting || userAnswerCount < 1}
+                    >
+                      {updating || submitting ? (
+                        <InlineSpinner size={40} />
+                      ) : (
+                        t('questionnaire/submit')
+                      )}
+                    </Button>
+                    <div {...styles.reset}>
+                      {userAnswerCount < 1 ? (
+                        t('questionnaire/invalid')
+                      ) : (
+                        <A href='#' onClick={this.handleReset}>
+                          {t('questionnaire/cancel')}
+                        </A>
+                      )}
+                    </div>
+                  </div>
+                )}
+              </div>
+            )
+          }}
+        />
+      </Frame>
+    )
+  }
+}
+
+export default compose(
+  withT,
+  withAuthorization(['supporter', 'editor'], 'showResults'),
+  withQuestionnaireMutation,
+  withQuestionnaireReset
+)(Page)
